feat(47): add check_tables option (#48)
Resolves #47

This pull request introduces a new feature to allow specifying tables
for sequential scan checks in the `pg_no_seqscan` PostgreSQL extension.
The main changes include updates to the README documentation, the
addition of new configuration settings, and enhancements to the codebase
to support the new functionality.

### Documentation Updates:
*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5R22-R32):
Added instructions and examples for the new `pg_no_seqscan.check_tables`
setting, which allows specifying tables to check for sequential scans.
[[1]](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5R22-R32)
[[2]](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5R50-R61)

### Configuration Enhancements:
*
[`src/guc.rs`](diffhunk://#diff-0ac1e7d5ab2f2480f77dd3e4222aec79ae39ebb6433336a5304b53b4a7e6e023R25-R27):
Introduced a new `GucSetting` for `pg_no_seqscan.check_tables` and
registered it in the `register_gucs` function.
[[1]](diffhunk://#diff-0ac1e7d5ab2f2480f77dd3e4222aec79ae39ebb6433336a5304b53b4a7e6e023R25-R27)
[[2]](diffhunk://#diff-0ac1e7d5ab2f2480f77dd3e4222aec79ae39ebb6433336a5304b53b4a7e6e023R47-R55)

### Codebase Enhancements:
*
[`src/hooks.rs`](diffhunk://#diff-899bcbd56696efca2a5e4e16bae25c24bc325b9849d561972ae53a19a899409cR106-R120):
Added methods to check if the `check_tables` option is set and to verify
if a table is included in the `check_tables` list. Updated the
sequential scan check logic to respect the new `check_tables` setting.
[[1]](diffhunk://#diff-899bcbd56696efca2a5e4e16bae25c24bc325b9849d561972ae53a19a899409cR106-R120)
[[2]](diffhunk://#diff-899bcbd56696efca2a5e4e16bae25c24bc325b9849d561972ae53a19a899409cL136-R155)

### Test Additions:
*
[`src/lib.rs`](diffhunk://#diff-b1a35a68f14e696205874893c07fd24fdb88882b47c23cc0e0c80a30c7d53759R161-R199):
Added new tests to verify the behavior of the `check_tables` setting and
its interaction with the `ignore_tables` setting.

---------

Co-authored-by: Emmanuel Quincerot <emmanuel.quincerot@doctolib.com>