fix: fix guc settings none (#52)
Making sure that string guc setting are never None.
Doing some refactoring as well to try have more homogeneous code.

Guc settings were set to default value as `None`.
In tests, we were testing:
* when these settings were None
* when these settings were containing a list
We missed the path where the settings are an empty string.

When the setting is loaded for the first time, the default value is
controlled by the code.

Then the only way to change the setting seems to be:
* `SET <setting_name> = <some_value>;` (null is not an allowed value)
* `SET <setting_name> TO DEFAULT;`
https://www.postgresql.org/docs/current/sql-set.html
* `RESET <setting_name>;`, equivalent to the previous command
https://www.postgresql.org/docs/current/sql-reset.html
* there is no `UNSET` command