Fix for different versions of cargo-pgrx using different RELKIND_SEQUENCE (#13)
Previously, the build of pgrx failed for certain versions of pg15 due to
the following error:

```
517.0    --> src/hooks.rs:178:43
517.0     |
517.0 178 |             (*relation.rd_rel).relkind == (pg_sys::RELKIND_SEQUENCE as i8)
^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ expected `u8`, found `i8`
```

This PR fixes the issue by updating `cargo-pgrx` to `0.15.0` that
defines a unified value for `RELKIND_SEQUENCE = c_char`.

The update of cargo-pgrx comes with changes in the signature of
`define_enum_guc`. This PR adapt the code to comply with those changes.

Other cosmetic changes:

- Adds comma_separated_list_contains function to check if a value is
contained in a comma separated list.