doc: avoid perf degradation because of jit (#32)
As pg_no_seqscan extension needs `enable_seqscan = off` the average cost
of queries increases strongly.
When working on partition tables with many partition or when join many
tables, the cost was big enough to let PG enable jit which is rarely a
good choice in terms of performance for local/test/ci database.

`jit_above_cost` was already set very high, but it's not enough for some
specific use cases.

Here we increase the recommendation.